requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pandas==2.1.3
python-multipart==0.0.6
openpyxl==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0